backend/requirements.txt
Flask
flask-cors
PyPDF2
pikepdf
pdf2image
Pillow
python-pptx
camelot-py[cv]
openpyxl
pdfplumber
pytesseract
python-docx
reportlab
PyMuPDF
beautifulsoup4
comtypes
pywin32; platform_system == 'Windows'

Flask
Flask-Cors
boto3
python-jose
PyPDF2
python-docx
Pillow
beautifulsoup4
camelot-py[cv]
comtypes
python-dotenv
PyMuPDF
pdf2image
pdfplumber
pikepdf
python-pptx
pytesseract
reportlab
requests
Werkzeug
